Move pdfjtIsBeforeVersion4 to FormImporter.
